% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_usable_files}
\alias{download_usable_files}
\title{Download Usable Files}
\usage{
download_usable_files(species_english_name, destination_path)
}
\arguments{
\item{species_english_name}{string: the English name of the species.}

\item{destination_path}{string: where to store the downloaded files. It
will be created if it doesn't already exist.}
}
\value{
a data frame listing the downloaded files
}
\description{
Downloads files from xeno-canto with usable licenses, i.e.,
not restriced by non-commercial or no-derivative clauses
}
\details{
\code{download_usable_files} uses the \code{warbleR} package to query the
\href{https://xeno-canto.org/}{xeno-canto} collection of bird sounds from
around the world. Xeno-canto licenes each file with one of the
\href{https://creativecommons.org/share-your-work/}{Creative Commons} licenses.

Since we want to analyze and resynthesize the sounds, and will be using
them in compositions that may be recorded and published, we can't use
files that have Creative Commons licenses with the "ND" (no derivative
works) or "NC" (non-commercial) attribute. That leaves us with "CC0"
(public domain), "CC-BY" (attribution), or "CC-BY-SA" (attribution and
share-alike).

It turns out that few of the xeno-canto files are usable by this
definition. For example, out of 424 recordings of the great horned owl,
only seven are usable. And most of the usable files are "CC-BY-SA", so
we use that license for output files distributed with \code{oiseaux}.
}
\examples{
\dontrun{
  owl_metadata <- oiseaux::download_usable_files(
    "Great Horned Owl", "~/owls"
  )
  loon_metadata <- oiseaux::download_usable_files(
    "Common Loon", "~/loons"
  )
}
}
